refactor: Restructure the source files and directories
